fix: Use parseFloat for checkInHours to support fractional hours
parseInt("0.083") returns 0, causing 5-min intervals to default to 72h.
Also enabled short interval options for all users for testing.

+++ frontend/app/dashboard/create-wizard/page.tsx
@@ -106,7 +106,7 @@ export default function CreateWizardPage() {
       const serverPayload = prepareServerPayload(
         encryptedSwitch,
         switchTitle,
-        parseInt(checkInHours),
+        parseFloat(checkInHours),
         validRecipients
       )
